修改了搜索的图片问题

+++ DiJieSheWeb/TourManagerDpt/UserManager.aspx.cs
@@ -84,6 +84,11 @@ public partial class TourManagerDpt_UserManager : basepageMgrDpt
         {
             Guid userid = Guid.Parse(e.CommandArgument.ToString());
             DJ_User_Gov user = blldj_user.GetGov_UserById(userid);
+            if ((Guid)CurrentUser.ProviderUserKey == userid)
+            {
+                ScriptManager.RegisterStartupScript(this, this.GetType(), "s", "alert('不得删除本人');", true);
+                return;
+            }
             blldj_user.DeleteGov_User(user);
             bind();
         }
